+++ index.php
@@ -8,7 +8,7 @@ require_once 'src/Entities/Album.php';
 require_once 'src/Services/DisplayThreeArtistsService.php';
 require_once 'src/Services/Choose3ArtistsService.php';
 require_once 'src/Models/SongsModel.php';
-require_once 'src/Services/PlayCountService.php';
+require_once 'src/Services/DisplayPlayButtonService.php';
 
 $db = DatabaseConnector::connect();
 
@@ -18,19 +18,6 @@ $artists = $artistsModel->getAllArtists();
 $albumsModel = new AlbumsModel($db);
 $album = new Album();
 $displayArtists = Choose3ArtistsService::choose3Artists($artists);
-
-//$songId = 3;
-//
-//$id = $_GET['playSong'];
-//
-//$playButton = new SongsModel($db);
-//$buttonTest = $playButton->getSongsByAlbum($songId);
-//
-//$playButton = UpdatePlayCount::playCount(4);
-//echo $playButton;
-//
-//$instantiate = new UpdatePlayCount($db);
-//$instantiate->updateDatabase($id);
 
 ?>
 <!DOCTYPE html>
