+++ src/Services/DisplayAlbumsSongs.php
@@ -3,10 +3,8 @@
 declare(strict_types=1);
 
 require_once 'src/Services/DatabaseConnector.php';
-require_once 'src/Services/PlayCountService.php';
+require_once 'src/Services/DisplayPlayButtonService.php';
 require_once 'artist.php';
-
-
 
 class DisplayAlbumsSongs {
 
@@ -28,7 +26,7 @@ class DisplayAlbumsSongs {
                         </div>
                         <div class='flex items-center justify-between w-24'>
                             <span class='text-slate-500'>{$songInAlbum->getLength()}</span>";
-                $albums .=  UpdatePlayCount::playCount($playId, $songInAlbum->getSongId());
+                $albums .=  DisplayPlayButtonService::playCount($playId, $songInAlbum->getSongId());
                 $albums .= "<a href='favourite.php?id=' class='hover:text-slate-500 hover:cursor-pointer'>
                                 <svg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 24 24' stroke-width='1.5' stroke='currentColor' class='size-6'>
                                     <path stroke-linecap='round' stroke-linejoin='round' d='M17.593 3.322c1.1.128 1.907 1.077 1.907 2.185V21L12 17.25 4.5 21V5.507c0-1.108.806-2.057 1.907-2.185a48.507 48.507 0 0 1 11.186 0Z' />
