+++ src/Services/DisplayArtistsAlbumsService.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-require_once 'src/Services/PlayCountService.php';
+require_once 'src/Services/DisplayPlayButtonService.php';
 
 class DisplayArtistsAlbumsService {
     public static function displayArtistsAlbums(array $artists, AlbumsModel $albumsModel): string
